app/requirements.txt
fastapi
aiofiles
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
